sentiflow/backend/requirements.txt
# Web Framework
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0

# Google Cloud (Python 3.12 compatible versions)
google-cloud-aiplatform>=1.43.0
google-cloud-storage>=2.14.0
google-cloud-bigquery>=3.14.0
google-cloud-pubsub>=2.18.0

# Elastic
elasticsearch==8.11.0
elastic-transport==8.11.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
requests==2.31.0

# Data Processing
PyPDF2==3.0.1
python-docx==1.1.0
